edit and remove finished!

+++ app/src/main/java/com/example/cookme/addItem.java
@@ -111,12 +111,12 @@ public class addItem extends AppCompatActivity {
 
                                 if(key!=-1)
                                 {
-                                    long currAmount = (long)ingredientList.get(key).getAmount();
-                                    long newAmount = currAmount + LAmount;
+                                    Double currAmount = ingredientList.get(key).getAmount();
+                                    Double newAmount = currAmount + LAmount;
 
                                     //update data syntax:
                                     //mDatabase.child("users").child(userId).child("username").setValue(name);
-                                    ingredientList.get(key).setAmount((Long)newAmount);
+                                    ingredientList.get(key).setAmount(newAmount);
 
                                     mRootRef.child("Inventory").child(groupId).setValue(ingredientList);
                                 }
